address review comments 2/2
