+++ homeassistant/components/kmtronic/switch.py
@@ -1,42 +1,17 @@
 """KMtronic Switch integration."""
-from datetime import timedelta
-import logging
-
-import aiohttp
-import async_timeout
 
 from homeassistant.components.switch import SwitchEntity
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import DATA_HOST, DATA_HUB, DOMAIN, MANUFACTURER
-
-_LOGGER = logging.getLogger(__name__)
+from .const import DATA_COORDINATOR, DATA_HOST, DATA_HUB, DOMAIN
 
 
 async def async_setup_entry(hass, entry, async_add_entities):
     """Config entry example."""
+    coordinator = hass.data[DOMAIN][entry.entry_id][DATA_COORDINATOR]
     hub = hass.data[DOMAIN][entry.entry_id][DATA_HUB]
     host = hass.data[DOMAIN][entry.entry_id][DATA_HOST]
     await hub.async_get_relays()
-
-    async def async_update_data():
-        try:
-            async with async_timeout.timeout(10):
-                await hub.async_update_relays()
-        except aiohttp.client_exceptions.ClientConnectorError as err:
-            raise UpdateFailed(f"Error communicating with API: {err}") from err
-
-    coordinator = DataUpdateCoordinator(
-        hass,
-        _LOGGER,
-        name=f"{MANUFACTURER} {hub.name}",
-        update_method=async_update_data,
-        update_interval=timedelta(seconds=30),
-    )
 
     async_add_entities(
         [KMtronicSwitch(coordinator, host, relay) for relay in hub.relays]
